+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Changed
+- Completely releasing sequences and modifier groups.
+
+### Added
+- Documented importance of mapping order.
 
 ## [Version 1.6.0] - 2021-08-29
 ### Added