+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
+## [Version 1.7.0] - 2021-12-13
 ### Changed
 - Removed implicitly mapping of modifier keys.
 - Matching begin of sequence when might-match failed.
@@ -11,6 +13,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Documented importance of mapping order.
+
+### Fixed
+- Made Linux keyboard initialization more robust.
 
 ## [Version 1.6.0] - 2021-08-29
 ### Added
@@ -63,7 +68,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Version 1.1.5] - 2020-05-09
 
-[Unreleased]: https://github.com/houmain/keymapper/compare/1.6.0...HEAD
+[Unreleased]: https://github.com/houmain/keymapper/compare/1.7.0...HEAD
+[Version 1.7.0]: https://github.com/houmain/keymapper/compare/1.6.0...1.7.0
 [Version 1.6.0]: https://github.com/houmain/keymapper/compare/1.5.0...1.6.0
 [Version 1.5.0]: https://github.com/houmain/keymapper/compare/1.4.0...1.5.0
 [Version 1.4.0]: https://github.com/houmain/keymapper/compare/1.3.0...1.4.0